requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
dj-database-url==0.5.0
Django==4.0.1
django-bootstrap4==21.2
django-braces==1.15.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
outcome==1.1.0
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wsproto==1.0.0
